mm/m again...
I've spent two hours and more, and i'm not the only one, to find all the correct info about the mm/s and mm/m backstory, starting by:

"Why add a *60 if i can calculate by myself the mm/m conversion, and why is it in mm/m? Am I missing something? "

A simple neat phrase will do the job and save an hour or two for newbies. And reference to the HOMING_BUMP_DIVISOR just to make sure.